Reduce allocations
The structure of getting a `PortableExecutableReference` meant that we
were essentially paying for `byte[]` twice. The first is generating the
`byte[]` for the resource and then when that was passed to
`CreateFromImage` it was converted to an `ImmutableArray<byte>` which
means there is yet another `byte[]` allocation. To avoid this the code
now just directly allocates the `ImmutableArray<byte>` and passes that
to `CreateFromImage`.

This can save up to 45MB in certain cases I was looking at in the Roslyn
tests.